Add MIPS tokenizer
The tokenizer identifies string literals, labels, dot-types, registers,
immediate values, instructions, addresses, and comments.  If two
instructions are placed on the same line, this will identify the first as
an instruction and the second as a label, however that's not too bad since
MIPS syntax only allows one instruction per line.