[Bugfix:Plagiarism] Bump base Docker image to Ubuntu 22.04 (#104)
### What is the current behavior?
Our Docker image is currently based on Ubuntu 20.04. This is problematic
because the main Lichen binary is compiled on the host machine and
copied into the Lichen container at runtime, which requires both
operating systems to be compatible. Eventually, we should do all of the
compilation at image build time instead...

### What is the new behavior?
The base Docker image has been bumped to Ubuntu 22.04. Due to a handful
of version incompatibilities and other miscellaneous Lichen bugs, I also
bumped clang to version 14 (the C++ tokenizer was broken), bumped python
to 3.9, and fixed a regression introduced in
https://github.com/Submitty/Submitty/pull/9630.

Further Python version increases will break the Java tokenizer. I will
make a separate PR to migrate away from `javac-parser`, which hasn't
been updated in 5 years now, and seems to be a dead project.